Memory optimization

+++ projects/game-engine-2/src/TienLenAi2.Core/Cards/Rank.cs
@@ -1,6 +1,6 @@
 namespace TienLenAi2.Core.Cards;
 
-public enum Rank : int
+public enum Rank : byte
 {
     Three = 3,
     Four = 4,
